requirements_test.txt
pytest>=7.4
pytest-asyncio
pytest-homeassistant-custom-component
